+++ backend/src/main/java/com/divipay/backend/repositories/sql/ExpenseSQLRepository.java
@@ -10,16 +10,16 @@ import org.springframework.stereotype.Repository;
 import com.divipay.backend.models.sql.ExpenseEntity;
 
 @Repository
-public interface ExpenseSQLRepository extends JpaRepository<ExpenseEntity, Integer> {
+public interface ExpenseSQLRepository extends JpaRepository<ExpenseEntity, Long> {
 
-    List<ExpenseEntity> findByEvent_Id(Integer eventId);
+    List<ExpenseEntity> findByEvent_Id(Long eventId);
 
-    List<ExpenseEntity> findByPayer_Id(Integer payerId);
+    List<ExpenseEntity> findByPayer_Id(Long payerId);
 
     // enlace a documento Mongo: buscar un gasto concreto
     Optional<ExpenseEntity> findByExternalDocId(String externalDocId);
 
-    Optional<ExpenseEntity> findByIdAndExternalDocId(Integer id, String externalDocId);
+    Optional<ExpenseEntity> findByIdAndExternalDocId(Long id, String externalDocId);
 
 
 }
